Made changes to summary

+++ FinalProject/FinalProject/Services/ISummaryService.cs
@@ -13,7 +13,7 @@ namespace FinalProject.Services
         void TrackMovie (string movieTitle, int id, string posterPath, MovieVotingHistoryDbContext db);
         void TrackIndecision (MovieVotingHistoryDbContext db);
         decimal GetAverageTimeSelected(MovieVotingHistoryDbContext db);
-        decimal GetFrequencySkipped(MovieVotingHistoryDbContext db);
+        int GetFrequencySkipped(MovieVotingHistoryDbContext db);
         decimal[] GetVotingAverage (MovieVotingHistoryDbContext db);
         //void TrackFood(MovieVotingHistoryDbContext db);
